Allow user to toggle sort
